+++ test/test-host.ts
@@ -5,7 +5,7 @@ import {
 } from "@typespec/compiler/testing";
 import type { EmitterOptions } from "../src/lib.js";
 
-const BaseTester = createTester(resolvePath(import.meta.dirname, "../.."), {
+export const BaseTester = createTester(resolvePath(import.meta.dirname, "../.."), {
   libraries: [
     "@massivescale/tsp-ts-client-models",
     "@typespec/http",